Add datetime-aware timeline merging and README guidance
Extend calculate_customer_timeline() so it can preserve datetime granularity in addition to legacy Date-based processing. The API now accepts gap units and difftime thresholds, auto-detects datetime inputs, and restores gap diagnostics in units that match the selected time class.

Update the C++ merge path to operate on numeric temporal values so POSIXct ranges remain POSIXct instead of being silently downgraded to Date. Regenerate the Rcpp bindings and Rd documentation to match the new interface.

Add tests for POSIXct timelines and character datetime auto-detection, and expand the README with clearer guidance on date ranges, datetime ranges, merge semantics, and practical timelining use cases.

+++ src/RcppExports.cpp
@@ -11,7 +11,7 @@ Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
 // merge_relationship_periods
-DataFrame merge_relationship_periods(DataFrame dtable, std::string id_column, std::string from_column, std::string to_column, CharacterVector characteristic_beg_columns, CharacterVector characteristic_end_columns, int gap_threshold, bool keep_all_periods);
+DataFrame merge_relationship_periods(DataFrame dtable, std::string id_column, std::string from_column, std::string to_column, CharacterVector characteristic_beg_columns, CharacterVector characteristic_end_columns, double gap_threshold, bool keep_all_periods);
 RcppExport SEXP _customerRelationship_merge_relationship_periods(SEXP dtableSEXP, SEXP id_columnSEXP, SEXP from_columnSEXP, SEXP to_columnSEXP, SEXP characteristic_beg_columnsSEXP, SEXP characteristic_end_columnsSEXP, SEXP gap_thresholdSEXP, SEXP keep_all_periodsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -22,7 +22,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< std::string >::type to_column(to_columnSEXP);
     Rcpp::traits::input_parameter< CharacterVector >::type characteristic_beg_columns(characteristic_beg_columnsSEXP);
     Rcpp::traits::input_parameter< CharacterVector >::type characteristic_end_columns(characteristic_end_columnsSEXP);
-    Rcpp::traits::input_parameter< int >::type gap_threshold(gap_thresholdSEXP);
+    Rcpp::traits::input_parameter< double >::type gap_threshold(gap_thresholdSEXP);
     Rcpp::traits::input_parameter< bool >::type keep_all_periods(keep_all_periodsSEXP);
     rcpp_result_gen = Rcpp::wrap(merge_relationship_periods(dtable, id_column, from_column, to_column, characteristic_beg_columns, characteristic_end_columns, gap_threshold, keep_all_periods));
     return rcpp_result_gen;
